+++ src/js/lib/tz-offset.js
@@ -1,3 +1,0 @@
-const secondsPerMinute = 60;
-
-export default () => new Date().getTimezoneOffset() * secondsPerMinute;